[SDK-109] Restructured the communication package in the SDK
The abstract communication has been restuctured to a package connectable and a package connector
