+++ njams-sdk/src/main/java/com/im/njams/sdk/Njams.java
@@ -31,11 +31,11 @@ import com.im.njams.sdk.common.Path;
 import com.im.njams.sdk.communication.CommunicationFactory;
 import com.im.njams.sdk.communication.InstructionListener;
 import com.im.njams.sdk.communication.NjamsSender;
-import com.im.njams.sdk.communication.Receiver;
+import com.im.njams.sdk.communication.connectable.Receiver;
 import com.im.njams.sdk.communication.ReplayHandler;
 import com.im.njams.sdk.communication.ReplayRequest;
 import com.im.njams.sdk.communication.ReplayResponse;
-import com.im.njams.sdk.communication.Sender;
+import com.im.njams.sdk.communication.connectable.Sender;
 import com.im.njams.sdk.configuration.Configuration;
 import com.im.njams.sdk.configuration.ConfigurationInstructionListener;
 import com.im.njams.sdk.configuration.ConfigurationProvider;
@@ -54,7 +54,7 @@ import com.im.njams.sdk.model.svg.ProcessDiagramFactory;
 import com.im.njams.sdk.serializer.Serializer;
 import com.im.njams.sdk.serializer.StringSerializer;
 import com.im.njams.sdk.settings.Settings;
-import com.im.njams.sdk.utils.StringUtils;
+
 import java.net.URL;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
