+++ njams-sdk/src/main/java/com/im/njams/sdk/communication/connector/Reconnector.java
@@ -1,7 +1,7 @@
 /*
  */
 
-package com.im.njams.sdk.communication.connection;
+package com.im.njams.sdk.communication.connector;
 
 import com.im.njams.sdk.common.NjamsSdkRuntimeException;
 import org.slf4j.Logger;
